+++ src/index.js
@@ -81,6 +81,10 @@ function createPushHandler(ctx) {
 const wrappedFetch = async (ctx, url, options = {}) => {
   const opts = { ...DEFAULT_FETCH_OPTIONS, ...options };
   let urlWithQuery;
+  // sanitze method name (#24)
+  if (typeof opts.method === 'string') {
+    opts.method = opts.method.toUpperCase();
+  }
   const lookupCache = CACHEABLE_METHODS.includes(opts.method)
     // respect cache mode (https://developer.mozilla.org/en-US/docs/Web/API/Request/cache)
     && !['no-store', 'reload'].includes(opts.cache);