+++ test/index.test.js
@@ -100,6 +100,21 @@ describe('Fetch Tests', () => {
     assert.deepEqual(jsonResponseBody.json, json);
   });
 
+  it('fetch sanitizes lowercase method names', async () => {
+    const method = 'post';
+    const json = { foo: 'bar' };
+    const resp = await fetch('https://httpbin.org/post', { method, json });
+    assert.equal(resp.status, 200);
+    assert.equal(resp.headers.get('content-type'), 'application/json');
+    const jsonResponseBody = await resp.json();
+    assert(jsonResponseBody !== null && typeof jsonResponseBody === 'object');
+    assert.deepEqual(jsonResponseBody.json, json);
+  });
+
+  it('fetch rejects on non-string method option', async () => {
+    assert.rejects(() => fetch('http://httpbin.org/status/200', { method: true }));
+  });
+
   it('fetch supports caching', async () => {
     const url = 'https://httpbin.org/cache/60'; // -> max-age=2 (seconds)
     // send initial request, priming cache
@@ -439,7 +454,6 @@ describe('Fetch Tests', () => {
 
   it('forcing HTTP/1(.1) works', async function test() {
     this.timeout(5000);
-
     // endpoint supporting http2 & http1
     const url = 'https://www.nghttp2.org/httpbin/status/200';
     // default context defaults to http2